sb_replace_files to sb_replace_files_log
replacing sb_replace_files with sb_replace_files_log and removing `sources` argument. The function seems to have a bug where it uploads the log file to SB (instead of saving it in the log file) and also pushes `sb_utils.R` to SB. This update seems to fix both of those issues.